YT-CPPGL-61: Prepare the GL module's documentation
- Updated the project license
- Added the Doxyfile and mkdocs.yml configuration files and a Makefile with documentation targets
- Added documentation comments to the public API elements of the GL module
- Prepared the tutorial pages and images for the GL module's documenation
- Defined a utility script for generating documentation pages for the library's concepts
- Renamed the `return_type` and `non_void_return_type` and related utility to use the `result` naming
- Modified the PFS template implementation to accept arbitrary* search nodes
- Renamed the `enqueue vertex predicate` types and arguments for algorithm templates to use the (search) node naming
- Fixed the Dijkstra's algorithm implementation to properly utilize custom PFS search nodes and properly handle distance comparison within the min heap
- Removed unused/redundant elements

+++ tests/source/gl/test_alg_coloring.cpp
@@ -82,6 +82,8 @@ TEST_CASE_TEMPLATE_DEFINE("bipartite coloring tests", TraitsType, traits_type_te
 
             fs::path coloring_file_path = data_path / "bicoloring_bipartite_graph_coloring.txt";
 
+            // Using uint8_t results in invalid value interpretation when reading from file
+            // Solution: the data is read as uint16_t and cast to uint8_t
             const auto coloring_values =
                 load_list<std::uint16_t>(sut.n_vertices(), coloring_file_path);
 
@@ -89,7 +91,9 @@ TEST_CASE_TEMPLATE_DEFINE("bipartite coloring tests", TraitsType, traits_type_te
                 coloring_values.begin(),
                 coloring_values.end(),
                 std::back_inserter(expected_coloring),
-                [](const std::uint16_t value) { return gl::bin_color_value{value}; }
+                [](const std::uint16_t value) {
+                    return gl::bin_color_value{static_cast<std::uint8_t>(value)};
+                }
             );
         }
